added shipping types

+++ migrations/20220813125602-order-status.js
@@ -17,7 +17,7 @@ exports.setup = function(options, seedLink) {
 exports.up = function(db) {
   return db.createTable('order_statuses', {
     id: {
-      type: 'int',
+      type: 'smallint',
       unsigned: true,
       primaryKey: true,
       autoIncrement: true
